requirements.txt
PyPDF2>=3.0.0
Pillow>=10.0.0
sentence-transformers>=2.2.2
torch>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
PyMuPDF>=1.23.0 